Create trip_fixes.m
Script to extract GPS data recorded during foraging trips identified during dive data analysis. It assumes that logger data are stored in a folder named like the site of the data to be analyised, and subfolders named like the birdID of each deployment. Within the birdID folder it expects to find the GPS data in a text file generated by TechnoSmart's X Manager.